requirements.txt
Django==2.1.4
pypiwin32==223
pytz==2018.7
pywin32==224
PyYAML==3.13
WMI==1.4.9
